submission/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Data processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# File handling and validation
python-multipart==0.0.6
pydantic[email]==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2



